// backend/src/entities/enums.ts
export enum Rol {
  ESTUDIANTE = 'ESTUDIANTE',
  EDUCADOR = 'EDUCADOR',
  ADMIN = 'ADMIN',
}

export enum EstadoMatricula {
  ACTIVA = 'ACTIVA',
  CANCELADA = 'CANCELADA',
  FINALIZADA = 'FINALIZADA',
}

export enum EstadoEscenario {
  DISPONIBLE = 'DISPONIBLE',
  RESERVADO = 'RESERVADO',
  EN_MANTENIMIENTO = 'EN_MANTENIMIENTO',
}

export enum TipoMantenimiento {
  PREVENTIVO = 'PREVENTIVO',
  CORRECTIVO = 'CORRECTIVO',
}